create new folder for registration and login, add jwt strategy
