+++ src/modules/auth/register/register.controller.ts
@@ -1,14 +1,13 @@
-import { Body, Controller, Post } from '@nestjs/common';
+import {Body, Controller, Post} from '@nestjs/common';
 import { RegisterService } from './register.service';
-import { CreateUserDto } from '../dto/create.user.dto';
-import { User } from '../users/users.entity';
+import { CreateUserDto } from '../../dto/create.user.dto';
+import { User } from '../../users/users.entity';
 import { ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger';
 
 @ApiTags('Регистрация')
 @Controller('auth')
 export class RegisterController {
   constructor(private readonly registerService: RegisterService) {}
-
   @ApiOperation({ summary: 'User creation' })
   @ApiResponse({ status: 200, type: User })
   @Post('/registration')
